imgencry/requirements.txt
fastapi
uvicorn
requests
sqlalchemy
python-multipart
tk
pycryptodome